test/requirements.txt
beartype
jaxlib
optax
pytest
scipy
tqdm
